requirements.txt
torch
transformers>=4.45.0
trl>=0.11.0
peft
accelerate
bitsandbytes
xformers
triton

unsloth[kaggle-new] @ git+https://github.com/unslothai/unsloth.git

pyyaml
pandas
datasets
scikit-learn
